+++ resources/cli-templates/controller-web.php
@@ -2,14 +2,14 @@
 
 namespace _namespace_\Controllers;
 
-use Bayfront\Bones\Controller;
+use Bayfront\Bones\Controllers\WebController;
 
 /**
  * _controller_name_ controller.
  *
  * Created for Bones v_bones_version_
  */
-class _controller_name_ extends Controller
+class _controller_name_ extends WebController
 {
 
     /**